gather state circuit test
