+++ zkevm-circuits/src/state_circuit/constraint_builder.rs
@@ -1,7 +1,7 @@
 use super::{
     lookups::Queries as LookupsQueries, multiple_precision_integer::Queries as MpiQueries,
-    random_linear_combination::Queries as RlcQueries, N_LIMBS_ACCOUNT_ADDRESS, N_LIMBS_ID,
-    N_LIMBS_RW_COUNTER,
+    random_linear_combination::Queries as RlcQueries,
+    param::*,
 };
 use crate::util::Expr;
 use crate::{
