+++ zkevm-circuits/src/state_circuit/multiple_precision_integer.rs
@@ -1,5 +1,5 @@
 use super::lookups;
-use super::{N_LIMBS_ACCOUNT_ADDRESS, N_LIMBS_RW_COUNTER};
+use super::param::*;
 use crate::util::Expr;
 use eth_types::{Address, Field};
 use halo2_proofs::{
