+++ zkevm-circuits/src/state_circuit/param.rs
@@ -1,0 +1,3 @@
+pub(super) const N_LIMBS_RW_COUNTER: usize = 2;
+pub(super) const N_LIMBS_ACCOUNT_ADDRESS: usize = 10;
+pub(super) const N_LIMBS_ID: usize = 2;
